Add files via upload
Java based application on Corona Virus that fetches real time data of total cases, deaths and recovered as well as the latest headlines. it has a chat bot (a sort of) and a self assessment tool. it also has a research hub which redirects to the research portals of different organizations. application is built using IntelliJ IDEA and Scene Builder. the only external library used is JSoup.